requirements.txt
django-markdownify
fontawesomefree